pyproject.toml
[tool.poetry]
name = "ddd-kafka-microservices"
version = "0.1.0"
description = ""
authors = ["kadimk"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^8.3.2"
faker = "^26.3.0"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
pymongo = {extras = ["srv"], version = "^4.8.0"}
motor = "^3.5.1"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.4.0"
punq = "^0.7.0"
pytest-asyncio = "^0.23.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
